create postcodeSearch component and input check
