+++ src/App.jsx
@@ -1,20 +1,28 @@
 import { useEffect, useState } from "react";
 import { getAreaData } from "./api";
+import PostcodeSearchForm from "./components/PostcodeSearchForm";
 
 import "./App.css";
 
 function App() {
   const [areas, setAreas] = useState([]);
-  const [postcodeTextInput, setPostcodeTextInput] = useState("");
+
   const [postcode, setPostcode] = useState("");
+
+  const [invalidInput, setInvalidInput] = useState(false);
 
   const load = async () => {
     try {
+      setInvalidInput(false);
       const areaData = await getAreaData(postcode);
 
       setAreas(areaData);
     } catch (error) {
-      window.alert("todo: fix app");
+      console.log(error);
+
+      setInvalidInput(true);
+      setAreas([]);
+      // window.alert("todo: fix app");
     }
   };
 
@@ -22,28 +30,13 @@ function App() {
     load();
   }, [postcode]);
 
-  const handleChange = (event) => {
-    setPostcodeTextInput(event.target.value);
-  };
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    setPostcode(postcodeTextInput);
-    setPostcodeTextInput("");
-  };
   return (
     <div className="App">
       <h1>Postcoders</h1>
-      <form className="postcodeSearchForm" onSubmit={handleSubmit}>
-        <input
-          type="text"
-          placeholder="Enter the outcode part of the postcode (eg: BB10 for 'BB10 2BQ)"
-          onChange={handleChange}
-          value={postcodeTextInput}
-          className="postcodeText"
-        />
-        <button className="searchBtn">Search</button>
-      </form>
+      <PostcodeSearchForm
+        setPostcode={setPostcode}
+        invalidInput={invalidInput}
+      />
       {areas.length ? (
         <h2>{`Areas for ${postcode}: ${areas.length}`}</h2>
       ) : null}
